chore: refact codes
