+++ models/configs.go
@@ -138,16 +138,18 @@ func ConfigsSetWithUname(ctx *ctx.Context, ckey, cval, uName string) error { //b
 
 	return err
 }
+
 func ConfigsGetFlashDutyAppKey(ctx *ctx.Context) (string, error) {
 	configs, err := ConfigsSelectByCkey(ctx, "flashduty_app_key")
 	if err != nil {
 		return "", err
 	}
-	if len(configs) == 0 {
-		return "", nil
+	if len(configs) == 0 || configs[0].Cval == "" {
+		return "", errors.New("flashduty_app_key is empty")
 	}
 	return configs[0].Cval, nil
 }
+
 func ConfigsSelectByCkey(ctx *ctx.Context, ckey string) ([]Configs, error) {
 	var objs []Configs
 	err := DB(ctx).Where("ckey=?", ckey).Find(&objs).Error
